+++ js/editing.js
@@ -1,0 +1,1 @@
+// Доп. задание: редактор фотографии
