+++ js/scrolling.js
@@ -1,0 +1,1 @@
+// Доп. задание: просмотр ленты
