+++ js/sending.js
@@ -1,0 +1,1 @@
+// Доп. задание: отправка данных на сервер
