+++ js/settings.js
@@ -1,0 +1,1 @@
+// Доп. задание: настройки ленты
