+++ js/util.js
@@ -1,0 +1,21 @@
+const getRandomInt = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (items) =>
+  items[getRandomInt(0, items.length - 1)];
+
+const createIdGenerator = () => {
+  let lastId = 0;
+
+  return () => {
+    lastId += 1;
+    return lastId;
+  };
+};
+
+export { getRandomInt, getRandomArrayElement, createIdGenerator };
